Complying with requested changes
