+++ core/java/iesi-rest-without-microservices/src/main/java/io/metadew/iesi/server/rest/scriptExecutionDto/tools/ExecutionInputParameterDto.java
@@ -7,7 +7,7 @@ import lombok.EqualsAndHashCode;
 @Data
 @AllArgsConstructor
 @EqualsAndHashCode
-public class ExecInputParameterDto {
+public class ExecutionInputParameterDto {
 
     private final String name;
     private final String value;
